Split out utils into seperate modules
