+++ dotcom-rendering/playwright/lib/islands.ts
@@ -1,18 +1,4 @@
-import type { BrowserContext, Cookie, Page } from '@playwright/test';
-
-// Playwright does not currently have a useful method for removing a single cookie
-// so this workaround is required
-const clearCookie = async (
-	context: BrowserContext,
-	cookieName: string,
-): Promise<void> => {
-	const cookies = await context.cookies();
-	const filteredCookies = cookies.filter(
-		(cookie: Cookie) => cookie.name !== cookieName,
-	);
-	await context.clearCookies();
-	await context.addCookies(filteredCookies);
-};
+import type { Page } from '@playwright/test';
 
 const waitForIsland = async (
 	page: Page,
@@ -36,4 +22,4 @@ const waitForIsland = async (
 	});
 };
 
-export { clearCookie, waitForIsland };
+export { waitForIsland };
